+++ src/main/java/lotto/domain/customer/Money.java
@@ -1,8 +1,8 @@
 package lotto.domain.customer;
 
-import lotto.exception.MoneyIsNegativeException;
-import lotto.exception.MoneyIsNotThousandUnitException;
-import lotto.exception.MoneyIsZeroException;
+import lotto.exception.money.MoneyIsNegativeException;
+import lotto.exception.money.MoneyIsNotThousandUnitException;
+import lotto.exception.money.MoneyIsZeroException;
 
 public class Money {
